Rel pxc 5.6.41

+++ mysql-test/mysql-test-run.pl
@@ -3417,6 +3417,12 @@ sub have_wsrep() {
   return defined $wsrep_on
 }
 
+sub wsrep_is_bootstrap_server($) {
+  my $mysqld= shift;
+  return $mysqld->if_exist('wsrep_cluster_address') &&
+    ($mysqld->value('wsrep_cluster_address') eq "gcomm://" ||
+     $mysqld->value('wsrep_cluster_address') eq "'gcomm://'");
+}
 
 sub check_wsrep_support() {
   if (have_wsrep())
@@ -3921,13 +3927,8 @@ sub wait_wsrep_ready($$) {
 
   for (my $loop= 1; $loop <= $loops; $loop++)
   {
-    if (run_query_output($mysqld, $query, $outfile) != 0)
-    {
-      $tinfo->{logfile}= "WSREP error while trying to determine node state";
-      return 0;
-    }
-
-    if (mtr_grab_file($outfile) =~ /^ON/)
+    if (run_query_output($mysqld, $query, $outfile) == 0 &&
+        mtr_grab_file($outfile) =~ /^ON/)
     {
       unlink($outfile);
       return 1;
@@ -5967,6 +5968,17 @@ sub start_servers($) {
       sleep_until_file_created("$datadir/auto.cnf", $opt_start_timeout,
                                $mysqld->{'proc'});
 
+      # If wsrep is on, we need to wait until the first
+      # server starts and bootstraps the cluster before
+      # starting other servers.
+      if (have_wsrep() && wsrep_is_bootstrap_server($mysqld))
+      {
+        mtr_verbose("WSREP waiting for first server to bootstrap cluster");
+        if (!wait_wsrep_ready($tinfo, $mysqld))
+        {
+          return 1;
+        }
+      }
     }
 
   }
